requirements.txt
djangorestframework
django-cors-headers
django-colorfield
pillow
djoser

django-cleanup
django-ckeditor
django-filter
pytils

selenium
selenium-requests

psycopg2
django-resized
drf-yasg




